+++ R/findInFiles.R
@@ -17,24 +17,26 @@
 #' @param excludeFoldersPattern a pattern; exclude from search the folders
 #'   which match this pattern
 #' @param root path to the root directory to search from
-#' @param output one of \code{"viewer"}, \code{"dataframe"} or
-#'   \code{"viewer+dataframe"}; see examples
+#' @param output one of \code{"viewer"}, \code{"tibble"} or
+#'   \code{"viewer+tibble"}; see examples
 #'
-#' @return A dataframe if \code{output="dataframe"}, otherwise a
+#' @return A tibble if \code{output="tibble"}, otherwise a
 #'   \code{htmlwidget} object.
 #'
 #' @import htmlwidgets
 #' @importFrom stringr str_split_fixed
 #' @importFrom crayon strip_style
+#' @importFrom tibble tibble
 #' @export
 #'
 #' @examples library(findInFiles)
 #' folder <- system.file("example", package = "findInFiles")
 #' findInFiles("R", "function", root = folder)
 #'
-#' findInFiles("R", "function", root = folder, output = "dataframe")
+#' findInFiles("R", "function", root = folder, output = "tibble")
 #'
-#' fif <- findInFiles("R", "function", root = folder, output = "viewer+dataframe")
+#' fif <- findInFiles("R", "function", root = folder, output = "viewer+tibble")
+#' FIF2tibble(fif)
 #' FIF2dataframe(fif)
 #' fif
 #'
@@ -52,7 +54,34 @@ findInFiles <- function(
     return(invisible(NULL))
   }
 
-  output <- match.arg(output, c("viewer", "dataframe", "viewer+dataframe"))
+  stopifnot(isString(ext))
+
+  if(isBinaryExtension(ext)){
+    stop(
+      sprintf("This file extension ('%s') is not allowed.", ext)
+    )
+  }
+
+  stopifnot(isString(output))
+  if(output == "dataframe"){
+    output <- "tibble"
+    message(
+      'The option `output="dataframe"` is deprecated; use `output="tibble"` ',
+      'instead.'
+    )
+  }
+  if(output == "viewer+dataframe"){
+    output <- "viewer+tibble"
+    message(
+      'The option `output="viewer+dataframe"` is deprecated; ',
+      'use `output="viewer+tibble"` instead.'
+    )
+  }
+
+  output <- match.arg(
+    output,
+    c("viewer", "tibble", "viewer+tibble")
+  )
 
   results <- grepInFiles(
     ext = ext, pattern = pattern, depth = depth,
@@ -62,30 +91,41 @@ findInFiles <- function(
     directory = root, output = output
   )
 
-  if(output %in% c("dataframe", "viewer+dataframe")){
-    if(output == "viewer+dataframe"){
+  if(output %in% c("tibble", "viewer+tibble")){
+    if(output == "viewer+tibble"){
       strippedResults <- strip_style(results)
     }else{
       strippedResults <- results
     }
     resultsMatrix <- stringr::str_split_fixed(strippedResults, ":", n = 3L)
     colnames(resultsMatrix) <- c("file", "line", "code")
-    resultsDF <- as.data.frame(resultsMatrix)
+    resultsDF <- as.data.frame(resultsMatrix, stringsAsFactors = FALSE)
     resultsDF[["line"]] <- as.integer(resultsDF[["line"]])
+    if(!is.null(results)){
+      opts <- attr(results, "options")
+      resultsDF[["code"]] <- do.call(
+        function(...){redifyVector(resultsDF[["code"]], ...)}, opts
+      )
+    }
+    resultsDF <- tibble(resultsDF)
     class(resultsDF) <- c(oldClass(resultsDF), "findInFiles")
-    if(output == "dataframe"){
+    if(output == "tibble"){
       return(resultsDF)
     }
   }
 
-  ansi <- paste0(results, collapse = "\n")
+  if(is.null(results)){
+    ansi <- "No results."
+  }else{
+    ansi <- paste0(results, collapse = "\n")
+  }
 
   # forward options using x
   if(output == "viewer"){
     x = list(
       ansi = ansi
     )
-  }else{ # viewer+dataframe
+  }else{ # viewer+tibble
     x = list(
       ansi = ansi,
       results = resultsDF
@@ -104,46 +144,76 @@ findInFiles <- function(
 
 }
 
-#' Output of `findInFiles` as a dataframe
+#' @title Output of `findInFiles` as a tibble
 #'
-#' Returns the results of \code{\link{findInFiles}} in a dataframe, when the
-#'   option \code{output = "viewer+dataframe"} is used.
+#' @description Returns the results of \code{\link{findInFiles}} in a
+#'   tibble, when the option \code{output = "viewer+tibble"} is used.
 #'
 #' @param fif the output of \code{\link{findInFiles}} used with the
-#'   option \code{output = "viewer+dataframe"}
+#'   option \code{output = "viewer+tibble"}
 #'
-#' @return The results of \code{\link{findInFiles}} in a dataframe.
+#' @return The results of \code{\link{findInFiles}} in a tibble.
 #' @export
 #'
 #' @examples folder <- system.file("example", package = "findInFiles")
-#' fif <- findInFiles("R", "function", root = folder, output = "viewer+dataframe")
-#' FIF2dataframe(fif)
+#' fif <- findInFiles("R", "function", root = folder, output = "viewer+tibble")
+#' FIF2tibble(fif)
 #' fif
-FIF2dataframe <- function(fif){
+FIF2tibble <- function(fif){
+  if(is.null(fif)){
+    return(NULL)
+  }
   if(is.data.frame(fif) && inherits(fif, "findInFiles")){
     return(fif)
   }
   if(!inherits(fif, c("findInFiles", "htmlwidget"))){
     stop(
-      "The `fif` argument is not a output of `findInFiles`.",
+      "The `fif` argument is not an output of `findInFiles`.",
       call. = TRUE
     )
   }
   output <- fif[["x"]][["results"]]
   if(is.null(output)){
     message(
-      'You did not set the option `output = "viewer+dataframe"`.'
+      'You did not set the option `output = "viewer+tibble"`.'
     )
     return(invisible(NULL))
   }
   output
 }
 
-
-#' Shiny bindings for \code{findInFiles}
+#' @title Output of `findInFiles` as a dataframe
 #'
-#' Output and render functions for using \code{findInFiles} within Shiny
-#' applications and interactive Rmd documents.
+#' @description Returns the results of \code{\link{findInFiles}} in a
+#'   dataframe, when the option \code{output = "viewer+tibble"} or
+#'   \code{output = "tibble"} is used.
+#'
+#' @param fif the output of \code{\link{findInFiles}} used with the
+#'   option \code{output = "viewer+tibble"} or \code{output = "tibble"}
+#'
+#' @return The results of \code{\link{findInFiles}} in a dataframe.
+#'
+#' @importFrom crayon strip_style
+#' @importFrom vctrs vec_data
+#' @export
+#'
+#' @examples folder <- system.file("example", package = "findInFiles")
+#' fif <- findInFiles("R", "function", root = folder, output = "viewer+tibble")
+#' FIF2dataframe(fif)
+#' fif
+FIF2dataframe <- function(fif){
+  tbl <- FIF2tibble(fif)
+  if(is.null(tbl)){
+    return(NULL)
+  }
+  tbl[["code"]] <- vec_data(strip_style(tbl[["code"]]))
+  as.data.frame(tbl)
+}
+
+#' @title Shiny bindings for \code{findInFiles}
+#'
+#' @description Output and render functions for using \code{findInFiles} within
+#'   Shiny applications and interactive Rmd documents.
 #'
 #' @param outputId output variable to read from
 #' @param width,height a valid CSS unit (like \code{"100\%"},
@@ -154,8 +224,14 @@ FIF2dataframe <- function(fif){
 #' @param quoted logical, whether \code{expr} is a quoted expression (with
 #'   \code{quote()})
 #'
+#' @return \code{FIFOutput} returns an output element that can be included in a
+#'   Shiny UI definition, and \code{renderFIF} returns a
+#'   \code{shiny.render.function} object that can be included in a Shiny server
+#'   definition.
+#'
 #' @name findInFiles-shiny
 #'
+#' @import htmlwidgets
 #' @export
 #'
 #' @examples library(findInFiles)